DMM-002 merge conflicts resolved, all tests working [PM,MP]
