+++ app/src/androidTest/java/com/team13/dealmymeal/MainActivityTest.kt
@@ -1,7 +1,9 @@
 package com.team13.dealmymeal
 
+import androidx.test.espresso.Espresso
 import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.ViewAssertion
+import androidx.test.espresso.action.ViewActions
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.assertion.ViewAssertions
 import androidx.test.espresso.matcher.ViewMatchers
@@ -9,6 +11,7 @@ import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.rules.ActivityScenarioRule
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import org.junit.Assert.*
+import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -18,6 +21,15 @@ class MainActivityTest{
 
     @get:Rule
     val rule = ActivityScenarioRule(MainActivity::class.java)
+
+
+    @Before
+    fun overview_clicked() {
+        Thread.sleep(500)
+        Espresso.onView(ViewMatchers.withId(R.id.navigation_addMeal)).perform(ViewActions.click())
+        Espresso.onView(ViewMatchers.withId(R.id.fragment_addMeal)).check(ViewAssertions.matches(ViewMatchers.isDisplayed()))
+        Thread.sleep(500)
+    }
 
     @Test
     fun buttonExists()
@@ -35,6 +47,7 @@ class MainActivityTest{
     @Test
     fun menuOpen()
     {
+        onView(withId(R.id.btnmenu)).perform(click())
         onView(withId(R.id.nav_bar_id)).check(ViewAssertions.matches(ViewMatchers.isDisplayed()))
 
     }
@@ -54,11 +67,5 @@ class MainActivityTest{
         onView(withId(R.id.nav_bar_id)).check(ViewAssertions.matches(ViewMatchers.withEffectiveVisibility(ViewMatchers.Visibility.INVISIBLE)))
     }
 
-    @Test
-    fun welcomeText()
-    {
-        onView(withId(R.id.btnmenu)).perform(click())
-        onView(withId(R.id.menu_welcometext)).check(ViewAssertions.matches(ViewMatchers.isDisplayed()))
-    }
 
 }
