+++ app/src/main/java/com/team13/dealmymeal/MainActivity.kt
@@ -23,42 +23,16 @@ class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(R.layout.main_activity)
         val navView: BottomNavigationView = findViewById(R.id.nav_view)
 
         val navController = findNavController(R.id.nav_host_fragment)
         // Passing each menu ID as a set of Ids because each
         // menu should be considered as top level destinations.
         val appBarConfiguration = AppBarConfiguration(setOf(
-                R.id.navigation_home, R.id.navigation_dashboard, R.id.navigation_overview))
+                R.id.navigation_addMeal, R.id.navigation_dashboard, R.id.navigation_overview))
         setupActionBarWithNavController(navController, appBarConfiguration)
         navView.setupWithNavController(navController)
-
-//start
-        val form_button: Button = findViewById(R.id.form_goAdd)
-
-
-
-        supportFragmentManager.setFragmentResultListener("requestKey", this) { requestKey, bundle ->
-            val result = bundle.getString("bundleKey")
-            form_button.visibility = View.VISIBLE
-        }
-
-        form_button.setOnClickListener() { v ->
-            val a = supportFragmentManager.beginTransaction()
-            val b = AddEntry()
-            a.replace(R.id.main, b)
-            a.addToBackStack(null)
-            form_button.visibility = View.GONE
-            a.commit()
-
-
-        }
-
-
-        if (supportFragmentManager.findFragmentById(R.id.fragment1)?.childFragmentManager?.isDestroyed == true) {
-            form_button.visibility = View.VISIBLE
-        }
 
 
     }
