--- app/src/main/java/com/team13/dealmymeal/ui/home/HomeFragment.kt
@@ -1,53 +1,0 @@
-package com.team13.dealmymeal.ui.home
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.ImageButton
-import android.widget.TextView
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.Observer
-import androidx.lifecycle.ViewModelProvider
-import com.google.android.material.navigation.NavigationView
-import com.team13.dealmymeal.R
-
-
-
-
-class HomeFragment : Fragment() {
-
-    private lateinit var homeViewModel: HomeViewModel
-
-    override fun onCreateView(
-            inflater: LayoutInflater,
-            container: ViewGroup?,
-            savedInstanceState: Bundle?
-    ): View? {
-        homeViewModel =
-                ViewModelProvider(this).get(HomeViewModel::class.java)
-        val root = inflater.inflate(R.layout.fragment_home, container, false)
-        val textView: TextView = root.findViewById(R.id.text_home)
-        homeViewModel.text.observe(viewLifecycleOwner, Observer {
-            textView.text = it
-        })
-
-        val btn = root.findViewById<ImageButton>(R.id.btnmenu)
-        val menu = root.findViewById<NavigationView>(R.id.nav_bar_id)
-        btn.setOnClickListener()
-        {
-            if(menu.visibility == View.VISIBLE)
-            {
-                menu.visibility = View.INVISIBLE
-            }
-            else
-            {
-                menu.visibility = View.VISIBLE
-            }
-
-        }
-
-        return root
-    }
-
-}
