+++ app/src/main/res/layout/fragment_addmeal.xml
@@ -1,40 +1,53 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    android:id="@+id/fragment1"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".AddEntry"
-    android:padding="10dp"
-    android:layout_margin="10dp"
-    android:background="@color/white">
+    tools:context=".ui.home.AddMealFragment"
+    android:paddingBottom="60dp"
+    android:id="@+id/fragment_addMeal"
+    >
 
-    <!-- TODO: Update blank fragment layout -->
+
+    <ImageButton
+        android:id="@+id/btnmenu"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:contentDescription="@string/MenuButtonHomeFragment"
+        android:src="@drawable/ic_burger_menu"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent"
+        android:paddingBottom="10dp"/>
+
+    <com.google.android.material.navigation.NavigationView
+        android:id="@+id/nav_bar_id"
+        android:layout_width="294dp"
+        android:layout_height="500dp"
+        android:visibility="invisible"
+        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toBottomOf="@+id/btnmenu" />
+
+
 
     <EditText
         android:id="@+id/form_edit"
         android:layout_width="match_parent"
         android:layout_height="50dp"
-        android:hint="Enter menu.."
+        android:hint="@string/enter_menu"
         app:layout_constraintBottom_toTopOf="@+id/check_veggie"
-        app:layout_constraintStart_toStartOf="parent" />
+        app:layout_constraintStart_toStartOf="parent"
+        android:inputType="text"
+        android:autofillHints="" />
 
     <Button
         android:id="@+id/form_save"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="save"
+        android:text="@string/save"
         app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent" />
-
-    <Button
-        android:id="@+id/form_cancel"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:text="cancel"
-        app:layout_constraintBottom_toBottomOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
 
     <RatingBar
@@ -42,24 +55,25 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:numStars="4"
-        app:layout_constraintBottom_toTopOf="@+id/form_cancel"
+        app:layout_constraintBottom_toTopOf="@+id/form_save"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent" />
 
     <TextView
         android:id="@+id/form_showEntry"
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:text="Add Entry..."
-        android:textSize="40dp"
-        tools:layout_editor_absoluteX="9dp"
-        tools:layout_editor_absoluteY="10dp" />
+        android:layout_width="400dp"
+        android:layout_height="wrap_content"
+        android:text="@string/add_entry"
+        android:textSize="40sp"
+        app:layout_constraintBottom_toTopOf="@+id/form_edit"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
     <CheckBox
         android:id="@+id/check_meat"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Meat"
+        android:text="@string/meat"
         app:layout_constraintBottom_toTopOf="@+id/form_ratingBar"
         app:layout_constraintStart_toStartOf="parent" />
 
@@ -67,7 +81,7 @@
         android:id="@+id/check_veggie"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Vegetarian"
+        android:text="@string/vegetarian"
 
         app:layout_constraintBottom_toTopOf="@+id/form_ratingBar"
         app:layout_constraintEnd_toStartOf="@+id/check_special"
@@ -77,9 +91,11 @@
         android:id="@+id/check_special"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Special"
+        android:text="@string/special"
         app:layout_constraintBottom_toTopOf="@+id/form_ratingBar"
         app:layout_constraintEnd_toEndOf="parent" />
+
+
