+++ app/src/main/res/menu/bottom_nav_menu.xml
@@ -2,9 +2,9 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
 
     <item
-        android:id="@+id/navigation_home"
-        android:icon="@drawable/ic_home_black_24dp"
-        android:title="@string/title_home" />
+        android:id="@+id/navigation_addMeal"
+        android:icon="@drawable/ic_baseline_add_24"
+        android:title="@string/add_meal" />
 
     <item
         android:id="@+id/navigation_dashboard"
