+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -6,10 +6,10 @@
     app:startDestination="@+id/navigation_dashboard">
 
     <fragment
-        android:id="@+id/navigation_home"
-        android:name="com.team13.dealmymeal.ui.home.HomeFragment"
+        android:id="@+id/navigation_addMeal"
+        android:name="com.team13.dealmymeal.ui.home.AddMealFragment"
         android:label="@string/title_home"
-        tools:layout="@layout/fragment_home" />
+        tools:layout="@layout/fragment_addmeal" />
 
     <fragment
         android:id="@+id/navigation_dashboard"
