+++ app/src/main/res/values/strings.xml
@@ -21,4 +21,13 @@
     <string name="title_activity_navigation_bar">MainActivity</string>
     <string name="delete">Delete</string>
     <string name="MenuButtonHomeFragment">MenuButtonHomeFragment</string>
+    <string name="add_meal">Add Meal</string>
+    <string name="enter_menu">Enter menu..</string>
+    <string name="save">save</string>
+    <string name="cancel">cancel</string>
+    <string name="add_entry">Add Entry...</string>
+    <string name="meat">Meat</string>
+    <string name="vegetarian">Vegetarian</string>
+    <string name="special">Special</string>
+    <string name="add">Add</string>
 </resources>
